Changed convention for library structure
- Now each class in the library is in its own file
- Each library has a "main.js" file which is supposed to be the main entry point into the library. It defines what is exported and how the library is supposed to be used.
  - For web components this file simply defines the component using `customElements`. It doesn't export anything and is only supposed to be imported for its side-effects.
